Added frontend UI

+++ backend/controllers/articleController.js
@@ -1,11 +1,18 @@
 const asyncHandler = require("express-async-handler")
 
 const Article = require("../model/articleModel")
+const User = require("../model/userModel")
 
 // Route = "/"
 // Method = GET
 const getArticles = asyncHandler(async (req, res) => {
     const articles = await Article.find()
+    res.status(200).json(articles)
+})
+
+
+const getArticlesByName = asyncHandler(async (req, res) => {
+    const articles = await Article.find({ user: req.user.id })
     res.status(200).json(articles)
 })
 
@@ -18,7 +25,8 @@ const postArticles = asyncHandler(async (req, res) => {
     }
     const article = await Article.create({
         header: req.body.header,
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(article)
@@ -31,6 +39,20 @@ const updateArticles = asyncHandler(async (req, res) => {
     if (!article) {
         res.status(400)
         throw new Error('Article new found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // If user does not exists
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(article.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     const updatedArticle = await Article.findByIdAndUpdate(
@@ -50,6 +72,20 @@ const deleteArticles = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Article new found')
     }
+    
+    const user = await User.findById(req.user.id)
+
+    // If user does not exists
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // Make sure the logged in user matches the goal user
+    if(article.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
 
     const deletedArticle = await Article.findByIdAndDelete(req.params.id)
 
@@ -61,5 +97,6 @@ module.exports = {
     getArticles,
     postArticles,
     updateArticles,
-    deleteArticles
+    deleteArticles,
+    getArticlesByName
 }
